Move routing server to dedicated package.
