+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/parametric/routing/RoutingWorkspaceService.java
@@ -32,7 +32,6 @@ import java.util.concurrent.ExecutorService;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.eclipse.lsp4j.ExecuteCommandParams;
 import org.rascalmpl.vscode.lsp.BaseWorkspaceService;
-import org.rascalmpl.vscode.lsp.LanguageServerRouter;
 import org.rascalmpl.vscode.lsp.util.concurrent.CompletableFutureUtils;
 
 /**
